+++ RELEASE_NOTES.md
@@ -7,4 +7,6 @@
 
 ## Buf Fixes
 
-- The "cache hit by database" sorting bug has been fixed in Dashboard.
+- The "cache hit by database" sorting bug has been fixed in Dashboard tab.
+- The "cache hit by database" sorting bug has been fixed in Tuning tab.
+- The "cache hit by table" sorting bug has been fixed in Tuning tab.